Set HashiCorp Terraform Extension Defaults
This updates the HashiCorp Extension settings to the updated naming convention.

The HashiCorp Extension [changed it's setting structure](https://github.com/hashicorp/vscode-terraform/blob/main/docs/settings-migration.md#settings-migration) for some settings in v2.24.0. Old settings are still honored for a period of time, but we should update this devcontainer with the correct names.