(GH-198) Add `extended` option to hugo feature (#240)
Prior to this change, the hugo feature only supported specifying the
version of Hugo to install. Hugo is available in two builds: standard,
which the feature already installs, and extended, which includes
functionality for post-processing CSS/SCSS/SASS and JavaScript.

This change adds a new `extended` option to the hugo feature, allowing
users to specify that they require the extended build of Hugo. It
defaults to `false` and installs the standard build of Hugo.

- Resolves #198